requirements.txt
ffmpeg-python==0.2.0
pyTelegramBotAPI==4.2.2
pytube==11.0.2
future==0.18.2